% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.letter.R
\name{a.letter}
\alias{a.letter}
\title{Quita las letras ñ y los espacios}
\usage{
a.letter(x)
}
\arguments{
\item{x}{un vector de caracteres}
}
\value{
un vector de caracteres
}
\description{
busca y quita las ñ por n y remplaza los espacion por -
}
