% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.numeric.factor.R
\name{as.numeric.factor}
\alias{as.numeric.factor}
\title{Convert a factor vector into a numeric vector}
\usage{
\method{as.numeric}{factor}(x)
}
\arguments{
\item{x}{a factor vector}
}
\value{
a numerical vector
}
\description{
trasnform a vector of factor into a numerical vector
}
\details{
name as.numeric.factor
rdname extras
}
\examples{

x<- sample( LETTERS[1:4], 1000, replace=TRUE, prob=c(0.1, 0.2, 0.65, 0.05) )
x<-as.factor(x)
x<-as.numeric.factor(x)
class(x)
}
