% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codifica.R
\name{codifica}
\alias{codifica}
\title{Imputa los codigos del INE a los muncipios}
\usage{
codifica(x, provincia)
}
\arguments{
\item{x}{un character con el nombre a buscar}

\item{provincia}{una provincia de Spain}
}
\description{
name codifica
}
