% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReport.R
\name{createReport}
\alias{createReport}
\title{Create a fact sheet for a specific project}
\usage{
createReport(
  om = NULL,
  salary = NULL,
  statusReport = NULL,
  cookie = NULL,
  id = NULL,
  theme = NULL,
  functionalGroup = NULL,
  section = NULL,
  division = NULL,
  region = NULL,
  status = "Approved",
  destdir = ".",
  path = "//dcnsbiona01a/BIODataSVC/IN/MSP/PowerBI-Projects/dataSPA/"
)
}
\arguments{
\item{om}{a data frame created by \code{getData(type='om')}}

\item{salary}{a data frame created by \code{getData(type='salary')}}

\item{statusReport}{a data frame created by \code{getData(type='statusReport')}}

\item{cookie}{a sessionid and csrftoken from a Department of
Fisheries and Oceans Canada (DFO) employee in the following
format: csrftoken=YOURTOKEN; sessionid=YOURSESSIONID}

\item{id}{the project_id from the Project Planning Tool (PPT)}

\item{theme}{theme classification of projects of either \verb{Mitigation of Ecosystem Stressors},
\verb{Marine Spatial Planning and Conservation}, \verb{Ecosystem Assessment and Climate Change},
\verb{Population Assessment and Recovery}, \code{Other}, \verb{Technology Development and Application}, and
\verb{Pacific Salmon}}

\item{functionalGroup}{classification of projects. Too see options do
\code{unique(om$function_group)}, where \code{om}is the output from \code{getData(type="om")}}

\item{section}{classification of projects referring to the sections at DFO. For
more details do \code{unique(om$section_display)}, where \code{om}is the output from
\code{getData(type="om")}}

\item{division}{classification of projects referring to the divisions at DFO. For
more details do \code{unique(om$section_display)}, where \code{om}is the output from
\code{getData(type="om")}}

\item{region}{parameter to specific specific region of either \code{Gulf}, \code{Maritimes},
\code{Pacific}, \code{Quebec}, or \verb{Ontario and Prairie}}

\item{status}{a character string indicating indicating which project
status to include (Approved,Reviewed, Draft, Submitted, Not Approved,
Recommended, and Cancelled). If NULL, all projects are included.
are included}

\item{destdir}{parameter indicating where to save the html report}

\item{path}{path to look for saved om_date. This must be the save as the path given when doing
\code{getData(type='om_date', keep=TRUE)}. This allows the report to print when the data was obtained
from the PPT. See Examples for more information.}
}
\value{
A fact sheet about the specified project id.
}
\description{
This function creates a fact sheet report for the
specified project id, theme, functional group, section,
or division. Each summary fact sheet is broken up into
different tabs: Summary, Research, Monitoring, Advice,
and Other. By default, this only looks at the Approved
projects.
}
\examples{
\dontrun{
# Example 1
om <- getData(type="om", cookie=cookie, keep=TRUE, path=".")
sal <- getData(type="salary", cookie=cookie, keep=TRUE, path=".")
omdate <- getData(type="om_date", cookie=cookie, keep=TRUE, path=".")
createReport(om=om, salary=sal, cookie=cookie, id=1093, path=".")
}
}
