% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.firm.a.R
\name{data.firm.a}
\alias{data.firm.a}
\title{Collect self-employment information at muncipality level}
\usage{
data.firm.a(provincia)
}
\arguments{
\item{provincia}{one of the 52 Spanish provinces. See \link{getbase.pob} for details.}
}
\value{
a data frame with the following variables:
}
\description{
Generate a data frame of self-employment information of a particular municipality
}
\details{
is an interative function, require the selection of a particular municipality.
}
\examples{
\dontrun{data.firm.a("Ceuta")}

}
