% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.firm.R
\name{data.firm}
\alias{data.firm}
\title{Collect firm information at muncipality level}
\usage{
data.firm(provincia)
}
\arguments{
\item{provincia}{one of the 52 Spanish provinces. See \link{getbase.pob} for details.}
}
\value{
a data frame with the following variables:
}
\description{
Generate a data frame of firms information of a particular municipality
}
\details{
is an interative function, require the selection of a particular municipality.
}
\examples{
\dontrun{data.firm("Ceuta")}

}
