% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empresa.a.R
\name{empresa.a}
\alias{empresa.a}
\title{Extract self-emplyment information}
\usage{
empresa.a(http)
}
\arguments{
\item{http}{caracther with the Axesor self-employment URL}
}
\value{
A data frame
}
\description{
Extract the self-employmet`s information from the Axesor webpage.
}
\details{
for variable contained in the data frame see XX
}
\examples{
empresa.a("https://autonomos.axesor.es//informe-de-autonomo/autonomo/Moussa-Gassa/5912075")

}
