% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empresa.R
\name{empresa}
\alias{empresa}
\title{Extract fimr`s information}
\usage{
empresa(http)
}
\arguments{
\item{http}{caracther with the Axesor firm URL}
}
\value{
A data frame
}
\description{
Extract the firm's information from the Axesor webpage.
}
\details{
for variable contained in the data frame see XX
}
\examples{
empresa("http://www.axesor.es/Informes-Empresas/5774886/ABRELATAS_COMUNICACION_SL.html")

}
