% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatSalaries.R
\name{formatSalaries}
\alias{formatSalaries}
\title{Format open salary information}
\usage{
formatSalaries(groups = NULL)
}
\arguments{
\item{groups}{character vector with the groups listed here: https://www.tbs-sct.canada.ca/pubs_pol/hrpubs/coll_agre/rates-taux-eng.asp Note that this function only supports select groups. Use "all" (default) to download all supported groups}
}
\value{
data frame with public service pay rates
}
\description{
This function is used to take the output of the get_salaries()
function in the TBSPayRates package and transform it into a
format that is compatible with the dataSPA package and HR
framework.
}
\examples{
\dontrun{
# getting all supported salaries
groups <- c("AI", "AO", "AV", "CS", "CX", "EC", "EL", "FB", "FI", "FS", "LP",
"NR", "PA", "RE", "RO", "SO", "SP", "TC", "TR", "UT")
df <- formatSalaries(groups=groups)
}
}
