% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{fundingPalette}
\alias{fundingPalette}
\title{A pallete for funding sources}
\description{
This was created so that each funding source is assigned a color.
This ensures funding sources are consistent with reports.
}
\examples{
\dontrun{
library(dataSPA)
data(fundingPalette)
fundingPalette
}
}
