% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.empresas.a.R
\name{get.empresas.a}
\alias{get.empresas.a}
\title{Genera URLs de Autonomos}
\usage{
get.empresas.a(http)
}
\arguments{
\item{http}{URL de pagina conteniendo la lista de URLs}
}
\value{
un vector de caracteres conteniendo las URLs
}
\description{
Crea las URLs de los Autonomos para extraer informacion
}
