% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get,empresa.R
\name{get.empresas}
\alias{get.empresas}
\title{Genera URLs de empresas}
\usage{
get.empresas(http)
}
\arguments{
\item{http}{URL de pagina conteniendo la lista de URLs}
}
\value{
un vector de caracteres conteniendo las URL
}
\description{
Crea las URLs de las empresas para extraer informacion
}
