% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbase.fen.R
\name{getbase.fen}
\alias{getbase.fen}
\title{Collect information of municipality population phenomena.}
\usage{
getbase.fen(year, provincia)
}
\arguments{
\item{year}{A numerical value between 1996 and the current year indicating the year of the required data.}

\item{provincia}{one of the 52 Spain’s province.}
}
\value{
a \code{xlsx} file saved in \code{data_poblacion} folder into the working directory wtih default name \code{fen_year_provincia}.
}
\description{
Download the population phenomena data of the province`s municipalities corresponding to the required.
}
\details{
You can use in capital lletter or not the names of spanish provincies. The full list of spanish provincies is:

"ALBACETE", "ALICANTE", "ALMERIA", "ARABA", "ASTURIAS", "AVILA", "BADAJOZ", "BALEARES", "BARCELONA", "BIZKAIA", "BURGOS", "CACERES", "CADIZ", "CANTABRIA", "CASTELLON", "CIUDAD REAL", "CORDOBA", "A CORUÑA", "CUENCA", "GIPUZKOA", "GIRONA", "GRANADA", "GUADALAJARA", "HUELVA", "HUESCA", "JAEN", "LEON", "LLEIDA", "LUGO", "MADRID", "MALAGA", "MURCIA", "NAVARRA", "OURENSE", "PALENCIA",  "LAS PALMAS", "PONTEVEDRA", "LA RIOJA", "SALAMANCA", "TENERIFE", "SEGOVIA", "SEVILLA","SORIA", "TARRAGONA", "TERUEL", "TOLEDO", "VALENCIA", "VALLADOLID", "ZAMORA", "ZARAGOZA", "CEUTA" and "MELILLA"
}
\examples{
getbase.fen(2005,"Madrid")

}
\seealso{
Other download functions: \code{\link{getbase.paro}},
  \code{\link{getbase.pob}}
}
