% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbase.paro.R
\name{getbase.paro}
\alias{getbase.paro}
\title{Collects information of unemployed at a municipality level.}
\usage{
getbase.paro(year, mes, provincia)
}
\arguments{
\item{year}{a numerical value between 2005 and the last available, which indicates the year of the required database.}

\item{mes}{oone of the 12 months –in Spanish language– indicating the month of the data collection.}

\item{provincia}{one of the 52 Spainish provinces.}
}
\value{
It is a \code{xlsx} file host in the \code{data_paro} folder which, in turn, is located inside the working directory called  \code{paro_MUNI_provincia_mmyy.xls}.
}
\description{
\code{getbase.paro} downloads data about the unemployed of the Spanish municipalities by province.
}
\details{
The month must be called as follows: "enero", "febrero", "marzo", "abril", "mayo", "junio", "julio", "agosto", "septiembre", "octubre", "noviembre" and "diciembre".

The names of the Spanish provinces may or may not be called. If yes, employ capital letters as follows:

"ALBACETE", "ALICANTE", "ALMERIA", "ARABA", "ASTURIAS", "AVILA", "BADAJOZ", " BALEARES", "BARCELONA", "BIZKAIA", "BURGOS", "CACERES", "CADIZ", "CANTABRIA", "CASTELLON", "CIUDAD REAL", "CORDOBA", "A CORUÑA", "CUENCA", "GIPUZKOA", "GIRONA", "GRANADA", "GUADALAJARA", "HUELVA", "HUESCA", "JAEN", "LEON", "LLEIDA", "LUGO", "MADRID", "MALAGA", "MURCIA", "NAVARRA", "OURENSE", "PALENCIA",  "LAS PALMAS", "PONTEVEDRA", "LA RIOJA", "SALAMANCA", "TENERIFE", "SEGOVIA", "SEVILLA","SORIA", "TARRAGONA", "TERUEL", "TOLEDO", "VALENCIA", "VALLADOLID", "ZAMORA", "ZARAGOZA", "CEUTA" and "MELILLA"
}
\examples{
getbase.paro(2005,"julio","Madrid")

}
\seealso{
Other download functions: \code{\link{getbase.fen}},
  \code{\link{getbase.pob}}
}
\concept{download functions}
