% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbase.pob.R
\name{getbase.pob}
\alias{getbase.pob}
\title{Collect information of municipality population.}
\usage{
getbase.pob(year, provincia, extr = FALSE, anual = FALSE)
}
\arguments{
\item{year}{A numerical value between 1996 and the current year indicating the year of the required data.}

\item{provincia}{one of the 52 Spain’s province.}

\item{extr}{Logical variable, is foreign population? FALSE is the default value.}

\item{anual}{logical variable, is data required by age? FALSE is the default value.}
}
\value{
a file saved in \code{data_poblacion} folder into the working directory wtih default name \code{pob_q_year_provincia.xls}.If \code{extr} is TRUE, the file will be saved as \code{pob_e_year_provincia.xls}. In case of \code{anual} is TRUE the file is saved by the name \code{pob_a_year_provincia.xls}
}
\description{
Download the population data of the province`s municipalities corresponding to the required.
}
\examples{
getbase.pob(2005,”Madrid”)

}
