% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get required data from the PPT API}
\usage{
getData(
  type = NULL,
  cookie = NULL,
  debug = 0,
  keep = FALSE,
  age = 7,
  path = "//dcnsbiona01a/BIODataSVC/IN/MSP/PowerBI-Projects/dataSPA/"
)
}
\arguments{
\item{type}{the type of data that is wished to be extracted
(either \code{om}, \code{om_date}, \code{salary}, or \code{salary_date}).
The types that end in \verb{_date} will return the date of creation
of a locally stored file.}

\item{cookie}{a sessionid and csrftoken from a Department of
Fisheries and Oceans Canada (DFO) employee in the following
format: csrftoken=YOURTOKEN; sessionid=YOURSESSIONID}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}

\item{keep}{logical value to optionally keep \code{om} data on the hard disk.
Default is FALSE and will not save data on the hard disk. Value of TRUE
will save \code{om} to disk in \code{path} unless there is an \code{om} file already in
\code{path} that was created more recently than \code{age} (number of days)}

\item{age}{maximum age in number of days that a file may be loaded.
Set to \code{0} to download new data every time.}

\item{path}{path to save file. Default is in the shared IN folder.}
}
\value{
dataframe
}
\description{
If \code{type} is \code{om}, this function extracts information from the
Project Planning Tool's (PPT) API to obtain information for
the O&M investment into projects. If type is \code{salary},
salary information pertaining to projects is obtained.
For type is \code{om_date} or \code{salary_date}, information about
the date the information was pulled is obtained.
}
\examples{
\dontrun{
library(dataSPA)
data(salaries)
cookie <- "csrftoken=YOURTOKEN; sessionid=YOURSESSIONID"
data <- getData(type="salary", cookie=cookie)
head(data,2)
}
\dontrun{
# Fix File ( ./dataSPA_om.rds ) does not exist error
 # Notice keep = TRUE saves the file, and the path argument says where to save it
om <- getData(type="om", cookie=cookie, keep=TRUE, path=".")
# Notice sal and om are saved in the same path
sal <- getData(type="salary", cookie=cookie, keep=TRUE, path=".")
 # Notice om_date is saved in the same path and done AFTER type=”om”
omdate <- getData(type="om_date", cookie=cookie, keep=TRUE, path=".")
}
}
\author{
Jaimie Harbin and Remi Daigle
}
