% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEBMpillars.R
\name{getEBMpillars}
\alias{getEBMpillars}
\title{Get the discriminating words used to identify EBM pillars or objectives.}
\usage{
getEBMpillars(
  file =
    "//dcnsbiona01a/BIODataSvc/IN/MSP/PowerBI-Projects/dataSPA/inputs/Stephenson_ebfm.R",
  n,
  ties = TRUE
)
}
\arguments{
\item{file}{the word or R file describing the EBM framework. Defaults to the IN folder}

\item{n}{highlights the n words with the highest term frequency–inverse document frequency (i.e. most important words that discriminate amongst pillars)}

\item{ties}{Should ties be kept together? The default, TRUE, may return more rows than you request. Use FALSE to ignore ties, and return the first n rows.}
}
\description{
Provides words and colours to be used in highlighting. Defaults to Stephenson et al. 2018 https://onlinelibrary.wiley.com/doi/full/10.1111/faf.12296
}
\author{
Remi Daigle and Jaimie Harbin
}
