% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth}
\alias{growth}
\title{Growth rate of an variable}
\usage{
growth(x, y, n)
}
\arguments{
\item{x}{a numeric vector corrspondin to \eqn{t}{t} time of a variable}

\item{y}{a numeric vector corrsponding to \eqn{t+n}{t+n} time in a varible}

\item{n}{period of time between x and y}
}
\value{
a growth rate, basically is the compute of an compound interest rate
}
\description{
Compute the groth rate of n periods od an x variable
}
