% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind.ev.R
\name{ind.ev}
\alias{ind.ev}
\title{Panel of demographic indexes at a municipality level for a period of time}
\usage{
ind.ev(inicio, fin, provincia, print = FALSE)
}
\arguments{
\item{inicio}{starting year of the panel, which must be higher than 1996.}

\item{fin}{last year of the panel.}

\item{provincia}{one of the 52 Spanish provinces.}

\item{print}{logical variable ‘do you need to print a output file with the results?’, being FALSE the default value}
}
\value{
It is a list conteining ten data frame.
}
\description{
\code{ind.ev} creates a list containing a panel of demographic indexes of Spain at the municipality level for a period of time from the years \code{inicio} to \code{fin}.
}
\details{
This function calculates ten demographical indexes:
\itemize{
\item Childhood index
	\item Youth index
	\item Third age index
	\item Dependence index
	\item Unemployment rate, both sexes
	\item Unemployment rate, males
	\item Unemployment rate, females
	\item Municipality average age, both sexes
	\item Municipality average age, males
\item Municipality average age, females
}
For a full description of the index, see the 2017 Socioeconomic Atlas of Extremadura
If \code{print} is set to \code{TRUE}, an \code{xlsx} file containing ten sheet -one  data frame per index- is saved into the folder \code{Outputs} which the name  \code{pob_ev_index_provincia_inicio-fin.xlsx}
}
\examples{
ind.ev(2005,2007,"Avila")

}
\references{
{
Junta de Extremadura (2017) Atlas Socieconómico de Extremadura 2017. Mérida (Spain).
\url{http://estadistica.gobex.es/web/guest/atlas-socieconomico-de-extremadura}
}
}
\concept{manipulation functions}
