% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista.empresa.a.R
\name{lista.empresa.a}
\alias{lista.empresa.a}
\title{Genrate Self-employment URLs list}
\usage{
lista.empresa.a(http)
}
\arguments{
\item{http}{character indicating the municipality’s URL in Axesor web site.}
}
\value{
a character vextor
}
\description{
Generate a list containing the self-employment`s URLs in Axesor web site
}
\examples{
\dontrun{lista.empresa.a("")}

}
