% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista.empresa.R
\name{lista.empresa}
\alias{lista.empresa}
\title{Genrate Firms URLs list}
\usage{
lista.empresa(http)
}
\arguments{
\item{http}{character indicating the municipality’s URL in Axesor web site.}
}
\value{
a character vextor
}
\description{
Generate a list containing the Firm’s URLs in Axesor web site
}
\examples{
lista.empresa("http://www.axesor.es/directorio-informacion-empresas/empresas-de-Alava/informacion-empresas-de-Alegria-Dulantzi/1")

}
