% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipio.a.R
\name{muncipio.a}
\alias{muncipio.a}
\alias{municipio.a}
\title{List of muncipalities URLs in Axesor self-employment data base}
\usage{
municipio.a(http)
}
\arguments{
\item{http}{character containing a province URL in Axesor self-employment data nase}
}
\value{
no me acuerdo
}
\description{
Generate a list muncipalities URLs in Axesor self-employment data base
}
\examples{
municipio("http://www.axesor.es/informe-de-autonomos/")

}
