% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipio.R
\name{muncipio}
\alias{muncipio}
\alias{municipio}
\title{List of muncipalities URLs in Axesor firms data base}
\usage{
municipio(http)
}
\arguments{
\item{http}{character containing a province URL in Axesor firms data nase}
}
\value{
no me acuerdo
}
\description{
Generate a list muncipalities URLs in Axesor firms data base
}
\examples{
municipio("http://www.axesor.es/directorio-informacion-empresas/empresas-de-Alava")

}
