% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.municipio.R
\name{nn.municipio}
\alias{nn.municipio}
\title{Nombre de municipio}
\usage{
nn.municipio(http)
}
\arguments{
\item{http}{una url de provincia en Axesor}
}
\value{
un vector de caractres, necesita ser mejorado para mostrar en pantalla.
}
\description{
Genera una lista numerada de los municipios
}
