% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num.firm.a.R
\name{num.firm.a}
\alias{num.firm.a}
\title{Collect the number of self-employment at muncipality level}
\usage{
num.firm.a(provincia)
}
\arguments{
\item{provincia}{one of the 52 Spanish provinces. See \link{getbase.pob} for details.}
}
\value{
A data frame
}
\description{
Generate a data frame with the number of selff-employment in each municipality
}
\examples{
\dontrun{num.firm.a("Araba")}

}
\seealso{
Other firms functions: \code{\link{data.firm.a}},
  \code{\link{data.firm}}, \code{\link{num.firm}}
}
