% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num.firm.R
\name{num.firm}
\alias{num.firm}
\title{Collect the number of firm at muncipality level}
\usage{
num.firm(provincia)
}
\arguments{
\item{provincia}{one of the 52 Spanish provinces. See \link{getbase.pob} for details.}
}
\value{
A data frame
}
\description{
Generate a data frame with the number of firms in each municipality
}
\examples{
\dontrun{num.firm("Araba")}

}
\seealso{
Other firms functions: \code{\link{data.firm.a}},
  \code{\link{data.firm}}, \code{\link{num.firm.a}}
}
