% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paro.R
\name{paro}
\alias{paro}
\title{import information of municipality unemploymet into R.}
\usage{
paro(year, mes = "julio", provincia)
}
\arguments{
\item{year}{A numerical value between 2005 and the current year indicating the year of the required data.}

\item{mes}{one of the 12 moths in Spanish indicating the moth when the data was collected.}

\item{provincia}{one of the 52 Spain’s province.}
}
\value{
data frame containing the fallowing variables.
- \code{cod} is the municipality identification number based in the INE codification.
- \code{Name} the municipality name.
- \code{Total unemployment} number of unemployed people in the municipality 
- \code{Total male unemployment} number of unemployes males in the municipality
- \code{Total female unemployment} number of unemployees females in the municipality
}
\description{
Import the work force data of the province`s municipalities corresponding to the required.
}
\examples{
#paro(2005,"julio","Madrid")

}
\seealso{
Other Loading functions: \code{\link{pob.a}},
  \code{\link{pob.e.tot}}, \code{\link{pob.e}},
  \code{\link{pob.fen}}, \code{\link{pob.h.tot}},
  \code{\link{pob.m.tot}}, \code{\link{pob.n.ev}},
  \code{\link{pob.n.tot}}, \code{\link{pob.q}},
  \code{\link{pob.tot}}
}
