% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parque.aut.R
\name{parque.aut}
\alias{parque.aut}
\title{Municipality vehicle fleet information}
\usage{
parque.aut(year, ca, provincia)
}
\arguments{
\item{year}{is numeric variable expressing the year of interest, must be higher than 2013}

\item{ca}{is a character indicating one of the 17 the autonomous community}

\item{provincia}{one of the 52 Spanish provincies}
}
\value{
A data frame containing the following variables: the municipality name, number and average age of vehicle fleet, type of vehicles (cars, vans, trucks, motorcycles, buses, etc) and some other variables related to the register of drivers, accidents and vehicle taxes.
}
\description{
Collect the municipality vehicle fleet information
}
\examples{
\dontrun{parque.aut(2014,"Ceuta y Melilla","Ceuta")}

}
