% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSPA.R
\name{plotSPA}
\alias{plotSPA}
\title{Plot an argoFloats Object}
\usage{
plotSPA(
  om = NULL,
  salary = NULL,
  which = NULL,
  id = NULL,
  funding = NULL,
  fundingChange = NULL,
  dataframe = FALSE,
  year = NULL,
  debug = 0
)
}
\arguments{
\item{om}{a data frame likely from \code{getData(type='om')}}

\item{salary}{a data frame likely from \code{getData(type='salary')}}

\item{which}{indicates which plot to plot (See Details). The
options to use include: \code{omBar}, \code{omPie}, \code{omAllocation},
\code{omAllocationGeneral}, \code{salaryBar}, \code{salaryAllocation},
\code{weekAllocation},\code{indeterminate}, or \code{predict}}

\item{id}{the project_id from the Project Planning Tool}

\item{funding}{a variable used when \code{which='predict'} used
to indicate which funding source will experience change in
the prediction}

\item{fundingChange}{a percentage used when \verb{which='predict}
used to indicate by how much the funding specified in
\code{funding} will change by}

\item{dataframe}{a Boolean indicating if the user wants the data frame
of what is being plotted returned}

\item{year}{a character indicating which year of the project to
plot. This is given in the format "YYYY-YYYY". If \code{year} is \code{NULL}
all years are plotted by default}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}
}
\value{
None (invisible NULL).
}
\description{
This function plots specific graphs (determined by the
\code{which} argument) using the data frames returned by
\code{\link[=getData]{getData()}}.
}
\details{
The various plot types are as follows:
\itemize{
\item For \code{which="omBar"}, a bar graph representing amount of money ($)
coded by funding type.
\item For \code{which="omPie"}, a pie chart representing amount of money ($)
coded by funding type
\item For \code{which="omAllocation"} a bar graph representing amount of money ($)
invested per O&M category for project years is plotted
\item For \code{which=omAllocationGeneral} a pie chat representing proportion of
funding per category type is plotted for year each of the specified project
\item For \code{which="salaryBar"} a bar graph representing amount of salary money ($)
using the median values of a specific job classification. Graph color
coded by funding type indicated by the legend.
\item For \code{which="salaryAllocation"} a bar graph representing amount of salary money ($)
using the median values of a specific job classification.
\item For \code{which="weekAllocation"} a bar graph representing time (weeks)
\item For \code{which="indeterminate"} a pie chart representing proportion of indeterminate vs
non-Indeterminate employees for project years is plotted
\item For \code{which="predict"}a line chart showing the trends for
changes in different funding scenarios
}
}
\examples{
\dontrun{
# Example 1: Plot Bar graph of O&M Allocations
library(dataSPA)
data <- getData(type="om", cookie=cookie)
plotSPA(om=data, which="omBar")

# Example 2: Plot a prediction of cost of station
# for NCP funding decreasing by 25\%, 50\%, 75\%, and 100\%
for the Snow Crab project

library(dataSPA)
data(salaries)
data <- getData(type='om', cookie=cookie)
data2 <- getData(type='salary', cookie=cookie)
plotSPA(om=data, salary=data2, which='predict', id=1093,
funding= rep("NCP (A-base)",4),
fundingChange=c(-0.25, -0.5, -0.75, -1))
}
}
\author{
Jaimie Harbin
}
