% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.e.R
\name{pob.e}
\alias{pob.e}
\title{Population grouped by nationality and sex}
\usage{
pob.e(year, provincia)
}
\arguments{
\item{year}{a numerical value from 1996 and the latest available, which indicates the year of the required database.}

\item{provincia}{one of the 52 Spainish provinces.}
}
\value{
It is a list containing a data frame with total population and population by sex. Each data frame contains the following variables: 
\itemize{
		\item \code{cod} is the municipality identification number based in the INE codification.
	\item \code{Name} is the municipality name.
		\item \code{Total} is the municipality population.
		\item Three variables containing information about population by age (younger than 16 years old, from 16 to 64 years old and older than 65 years old), which are, in turn, grouped by total population, foreign population and national population.
}
}
\description{
\code{pob.e} imports into R the municipality population grouped by nationality and sexa
}
\examples{
pob.e(2016,"Madrid")

}
\seealso{
Other Loading functions: \code{\link{paro}},
  \code{\link{pob.a}}, \code{\link{pob.e.tot}},
  \code{\link{pob.fen}}, \code{\link{pob.h.tot}},
  \code{\link{pob.m.tot}}, \code{\link{pob.n.ev}},
  \code{\link{pob.n.tot}}, \code{\link{pob.q}},
  \code{\link{pob.tot}}
}
\concept{Loading functions}
