% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.fen.ev.R
\name{pob.fen.ev}
\alias{pob.fen.ev}
\title{Population phenomena evolution}
\usage{
pob.fen.ev(inicio, fin, provincia, print = FALSE)
}
\arguments{
\item{inicio}{starting year of the panel. Must be higher than 1996.}

\item{fin}{last year of the panel.}

\item{provincia}{one of the 52 Spanish provinces.}

\item{print}{Logical variable, do you need print a output file with the results?. Default value is FALSE}
}
\value{
a list conteining two data frame one of birth population at municpality level (Nacimientos) and other with death population at muncipality level (Falleciemientos)
}
\description{
Create a list containing the provincie population phenomena evolution at municipality level between the \code{inicio} and \code{fin} years.
}
\details{
If \code{print} is \code{TRUE}, a \code{xlsx} file containing two sheet, one per each data frame, is saved into the folder \code{Outputs} by the name:
\code{fen_evol_total_provincia_inicio-fin.xlsx}
}
\examples{
pob.fen.ev(2005,2007,"Avila")

}
\seealso{
Other Manipulate functions: \code{\link{ind.ev}},
  \code{\link{pob.e.ev}}, \code{\link{pob.ev}},
  \code{\link{pob.h.ev}}, \code{\link{pob.ind.p}},
  \code{\link{pob.ind}}, \code{\link{pob.m.ev}},
  \code{\link{pob.n.ev}}
}
