% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.h.ev.R
\name{pob.h.ev}
\alias{pob.h.ev}
\title{Total male population evolution}
\usage{
pob.h.ev(inicio, fin, provincia, print = FALSE)
}
\arguments{
\item{inicio}{starting year of the panel. Must be higher than 1996.}

\item{fin}{last year of the panel.}

\item{provincia}{one of the 52 Spanish provinces.}

\item{print}{Logical variable, do you need print a output file with the results?. Default value is FALSE}
}
\value{
a data frame
}
\description{
Create a data drame containing the provincie total male population evolution at municipality level between the \code{inicio} and \code{fin} years.
}
\details{
If \code{print} is \code{TRUE}, a \code{xlsx} file containing the data frame is saved into the folder \code{Outputs} by the name:
\code{pob_male_ev_provincia_inicio-fin.xlsx}
}
\examples{
pob.h.ev(2005,2007,"Avila")

}
