% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.h.tot.R
\name{pob.h.tot}
\alias{pob.h.tot}
\title{Total Male Population  data}
\usage{
pob.h.tot(year, provincia)
}
\arguments{
\item{year}{A numerical value between 1996 and the current year indicating the year of the required data.}

\item{provincia}{one of the 52 Spain`s province.}
}
\value{
a  data frame conatining the municipality total male population
}
\description{
import into R the total male municpality population
}
\examples{
pob.h.tot(2016,"Madrid")

}
