% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.ind.p.R
\name{pob.ind.p}
\alias{pob.ind.p}
\title{Calculate some demographic indexes}
\usage{
pob.ind.p(year, provincia, print = FALSE)
}
\arguments{
\item{year}{a numerical value from 1996 and the latest available year, which indicates the year of the required database.}

\item{provincia}{one of the 52 Spanish provinces.}

\item{print}{logical variable ‘do you need to print a output file with the results?’ being FALSE the default value.}
}
\value{
A data frame contains the spatial units, in rows, and the demographic index computed in  columns.
}
\description{
\code{pob.ind.p} computes a set of demographic indexes
}
\details{
This function calculates ten demographical indexes:
\itemize{
\item Childhood index
	\item Youth index
	\item Third age index
	\item Dependence index
	\item Unemployment rate, both sexes
	\item Unemployment rate, males
	\item Unemployment rate, females
	\item Municipality average age, both sexes
	\item Municipality average age, males
\item Municipality average age, females
}
For a full description of the index, see the 2017 Socioeconomic Atlas of Extremadura 

If \code{print} is set to \code{TRUE}, an \code{xlsx} file containing ten sheet -one  data frame per index- is saved into the folder \code{Outputs} which the name  \code{pob_index_provincia_year.xlsx}
The principal difference between this function and \link{pob.ind} is the precision of the calculation of the municipality average age. Since the year 2011, this function uses the one-year age variable for the computation instead of the five-year age approximation used in \code{pob.ind}. For more details, see the 2017 Socioeconomic Atlas of Extremadura.
}
\examples{
pob.ind.p(2012,"Madrid")

}
\references{
{
Junta de Extremadura (2017) Atlas Socieconómico de Extremadura 2017. Mérida (Spain).
\url{http://estadistica.gobex.es/web/guest/atlas-socieconomico-de-extremadura}
}
}
\seealso{
Other Manipulate functions: \code{\link{pob.e.ev}},
  \code{\link{pob.ev}}, \code{\link{pob.fen.ev}},
  \code{\link{pob.h.ev}}, \code{\link{pob.ind}},
  \code{\link{pob.m.ev}}, \code{\link{pob.n.ev}}
}
\concept{Manipulate functions}
