% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.ind.R
\name{pob.ind}
\alias{pob.ind}
\title{Calculate the population indexs}
\usage{
pob.ind(year, provincia, print = FALSE)
}
\arguments{
\item{year}{A numerical value between 1996 and the current year indicating the year of the required data.}

\item{provincia}{one of the 52 Spanish provinces.}

\item{print}{Logical variable, do you need print a output file with the results?. Default value is FALSE}
}
\value{
A data frame contains the spatial units on the brows and the index coputed in the columns.
}
\description{
Compute a set of population index
}
\details{
in this function 10 index are comptuted:
\itemize{
\item Childhood index
	\item Youthfulness index
	\item Eld index
	\item Dependence index
	\item Total unemployment rate
	\item Male unemployment rate
	\item Femal unemployment rate
	\item Total Population average age
	\item Male population average age
\item Female population average age
}
For a full description of the index, see the 2017 Socioeconomic Atlas of Extremadura
}
\examples{
pob.ind(2012,"Madrid")

}
\references{
{
Junta de Extremadura (2017) Atlas Socieconómico de Extremadura 2017. Mérida (Spain).
\url{http://estadistica.gobex.es/web/guest/atlas-socieconomico-de-extremadura}
}
}
\seealso{
Other Manipulate functions: \code{\link{ind.ev}},
  \code{\link{pob.e.ev}}, \code{\link{pob.ev}},
  \code{\link{pob.fen.ev}}, \code{\link{pob.h.ev}},
  \code{\link{pob.ind.p}}, \code{\link{pob.m.ev}},
  \code{\link{pob.n.ev}}
}
