% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.m.ev.R
\name{pob.m.ev}
\alias{pob.m.ev}
\title{Panel of the female population at the municipality level for a period of time}
\usage{
pob.m.ev(inicio, fin, provincia, print = FALSE)
}
\arguments{
\item{inicio}{starting year of the panel, which must be higher than 1996..}

\item{fin}{last year of the panel.}

\item{provincia}{one of the 52 Spanish provinces.}

\item{print}{logical variable ‘do you need to print a output file with the results?’ being FALSE the default value.}
}
\value{
It is a data frame
}
\description{
pov.m.ev create a data frame containing the panel of the female population of Spain at the municipality level for a period of time from the  \code{inicio} to \code{fin}.
}
\details{
If \code{print} is set to \code{TRUE}, a \code{xlsx} file containing the data frame is saved in the folder \code{Outputs} which name \code{pob_female_ev_provincia_inicio-fin.xlsx}
}
\examples{
pob.m.ev(2005,2007,"Avila")

}
\seealso{
Other Manipulate functions: \code{\link{pob.e.ev}},
  \code{\link{pob.ev}}, \code{\link{pob.fen.ev}},
  \code{\link{pob.h.ev}}, \code{\link{pob.ind.p}},
  \code{\link{pob.ind}}, \code{\link{pob.n.ev}}
}
\concept{Manipulate functions}
