% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.m.tot.R
\name{pob.m.tot}
\alias{pob.m.tot}
\title{Total Female Population  data}
\usage{
pob.m.tot(year, provincia)
}
\arguments{
\item{year}{A numerical value between 1996 and the current year indicating the year of the required data.}

\item{provincia}{one of the 52 Spain`s province.}
}
\value{
a  data frame conatining the municipality total female population
}
\description{
import into R the total female municpality population
}
