% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.n.ev.R
\name{pob.n.ev}
\alias{pob.n.ev}
\title{Total national population evolution}
\usage{
pob.n.ev(inicio, fin, provincia, print = FALSE)
}
\arguments{
\item{inicio}{starting year of the panel. Must be higher than 1996.}

\item{fin}{last year of the panel.}

\item{provincia}{one of the 52 Spanish provinces.}

\item{print}{Logical variable, do you need print a output file with the results?. Default value is FALSE}
}
\value{
a data frame
}
\description{
Create a data drame containing the provincie total national population evolution at municipality level between the \code{inicio} and \code{fin} years.
}
\details{
If \code{print} is \code{TRUE}, a \code{xlsx} file containing the data frame is saved into the folder \code{Outputs} by the name:
\code{pob_national_ev_provincia_inicio-fin.xlsx}
}
\examples{
pob.n.ev(2005,2007,"Avila")

}
\seealso{
Other Loading functions: \code{\link{paro}},
  \code{\link{pob.a}}, \code{\link{pob.e.tot}},
  \code{\link{pob.e}}, \code{\link{pob.fen}},
  \code{\link{pob.h.tot}}, \code{\link{pob.m.tot}},
  \code{\link{pob.n.tot}}, \code{\link{pob.q}},
  \code{\link{pob.tot}}

Other Manipulate functions: \code{\link{ind.ev}},
  \code{\link{pob.e.ev}}, \code{\link{pob.ev}},
  \code{\link{pob.fen.ev}}, \code{\link{pob.h.ev}},
  \code{\link{pob.ind.p}}, \code{\link{pob.ind}},
  \code{\link{pob.m.ev}}
}
