% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.n.tot.R
\name{pob.n.tot}
\alias{pob.n.tot}
\title{Total National Population  data}
\usage{
pob.n.tot(year, provincia)
}
\arguments{
\item{year}{A numerical value between 1996 and the current year indicating the year of the required data.}

\item{provincia}{one of the 52 Spain`s province.}
}
\value{
a  data frame conatining the municipality total national population
}
\description{
import into R the total national municpality population
}
