% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.q.R
\name{pob.q}
\alias{pob.q}
\title{Population by quinquennials age group data}
\usage{
pob.q(year, provincia)
}
\arguments{
\item{year}{A numerical value between 1996 and the current year indicating the year of the required data.}

\item{provincia}{one of the 52 Spain`s province.}
}
\value{
a list containing a total population data frame and the population grouped by sex. Each data frame contains the following variables: 
- \code{cod} is the municipality identification number based in the INE codification.
- \code{Name} the municipality name.
- \code{Total} total municipality population
- 21 variables containing the population by quinquennials age group
}
\description{
import into R the  population by quinquennials age group data
}
\examples{
pob.q(2016,"Caceres")

}
