% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.tot.R
\name{pob.tot}
\alias{pob.tot}
\title{Total Population  data}
\usage{
pob.tot(year, provincia)
}
\arguments{
\item{year}{A numerical value between 1996 and the current year indicating the year of the required data.}

\item{provincia}{one of the 52 Spain`s province.}
}
\value{
a  data frame conatining the municipality total population
}
\description{
import into R the total municpality population
}
