% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{salaries}
\alias{salaries}
\title{A pay scale spread sheet}
\description{
This includes salary for different levels of different
positions in different years. This information is used
to calculate median salaries spent on projects to ensure
confidentiality.
}
\examples{
library(dataSPA)
data(salaries)
head(salaries,5)
}
