% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{salary}
\alias{salary}
\title{An example project including salary information}
\description{
This includes a sample of salary data obtained from an
imaginary project. The purpose of this is for the user
to get familiar with the structure of the data,
and to allow the developers to validate their calculations
in a suite of tests.
}
\examples{
\dontrun{
library(dataSPA)
data(salary)
plotSPA(salary=salary, which="salaryAllocation", id=1234)
}
}
