% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCap.R
\name{simpleCap}
\alias{simpleCap}
\title{Coloca la primer letra en mayuscula}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{un vector de caracteres}
}
\value{
un vector de caracteres
}
\description{
coloca la primer letra en mayuscula de un vector de caracteres
}
