% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapText.R
\name{wrapText}
\alias{wrapText}
\title{Create new line for long string
This function identifies which strings have charaters greater than
the specified \code{nchar} value. If the string is greater than the
specified \code{nchar}, it checks if the nchar(th) character is a space.
If it is, a new line is created. If it isn't, it looks for
the next space and creates a new line there. This can be used
in legends when funding source names are longer than desired.}
\usage{
wrapText(string, nchar = 20, debug = 0)
}
\arguments{
\item{string}{a character string or list of character strings}

\item{nchar}{the number indicating the position to create a
new line in a string}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}
}
\description{
Create new line for long string
This function identifies which strings have charaters greater than
the specified \code{nchar} value. If the string is greater than the
specified \code{nchar}, it checks if the nchar(th) character is a space.
If it is, a new line is created. If it isn't, it looks for
the next space and creates a new line there. This can be used
in legends when funding source names are longer than desired.
}
\examples{
\dontrun{
 # 1. Try a string less than 20 characters
 cat(wrapText("This is a test", debug=1))

 # 2. Wrap when 20th character is a space
 cat(wrapText("1234567890123456789 1234"))

 # 3. Wrap text when 20th character is not a space
 # and there is a space after
 cat(wrapText("ABCDEFGH IJKLMONP QRS TU W ZYZ"))

# 4. wrap text when 20th character not space
and there is no space after
cat(wrapText("ABCDEF GHIJHLMOPDSDABBBBAAAASASASJH"))
}
}
