% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReport.R
\name{createReport}
\alias{createReport}
\title{Create a fact sheet for a specific project}
\usage{
createReport(
  om = NULL,
  salary = NULL,
  id = NULL,
  Rmdpath = "./R/",
  destdir = "."
)
}
\arguments{
\item{om}{a data frame created by \code{getData(type='om')}}

\item{salary}{a data frame created by \code{getData(type='salary')}}

\item{id}{the project_id from the Project Planning Tool (PPT)}

\item{Rmdpath}{the path to the template for the html output}

\item{destdir}{parameter indicating where to save the html report}
}
\value{
A html fact sheet about the specified project id.
}
\description{
This function creates a fact sheet report for the
specified project id.
}
\examples{
\dontrun{
library(dataSPA)
data(salaries)
data <- getData(type="om",cookie=cookie)
data2 <- getData(type="salary", cookie=cookie, salaries=salaries)
createReport(om=data, salary=data2, Rmdpath="./R")
}
}
