% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReport.R
\name{createReport}
\alias{createReport}
\title{Create a fact sheet for a specific project}
\usage{
createReport(
  om = NULL,
  salary = NULL,
  cookie = NULL,
  id = NULL,
  destdir = ".",
  output = "html"
)
}
\arguments{
\item{om}{a data frame created by \code{getData(type='om')}}

\item{salary}{a data frame created by \code{getData(type='salary')}}

\item{cookie}{a sessionid and csrftoken from a Department of
Fisheries and Oceans Canada (DFO) employee in the following
format: csrftoken=YOURTOKEN; sessionid=YOURSESSIONID}

\item{id}{the project_id from the Project Planning Tool (PPT)}

\item{destdir}{parameter indicating where to save the html report}

\item{output}{a character string of either \code{html}, \code{pdf}, or \code{word}
indicating which type of file output to create.}
}
\value{
A fact sheet about the specified project id.
}
\description{
This function creates a fact sheet report for the
specified project id.
}
\examples{
\dontrun{
library(dataSPA)
data <- getData(type="om",cookie=cookie)
data2 <- getData(type="salary", cookie=cookie)
createReport(om=data, salary=data2, id=1234, cookie=cookie)
}
}
