% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.firm.a.R
\name{data.firm.a}
\alias{data.firm.a}
\title{Collects information of self-employed at a municipality level}
\usage{
data.firm.a(provincia)
}
\arguments{
\item{provincia}{one of the 52 Spanish provinces. See \link{getbase.pob} for details.}
}
\value{
A data frame containing the following variables for each self-employment: location (province, municipality, address), company characteristics (name, birth, legal form, social object), economic activity codes and self-employment URL
}
\description{
\code{data.firm.a} generates a data frame of self-employment information of a particular municipality
}
\details{
It is an interactive function, which requires the selection of a particular municipality.
}
\examples{
\dontrun{data.firm.a("Ceuta")}

}
\seealso{
Other firm functions: \code{\link{data.firm}},
  \code{\link{num.firm.a}}, \code{\link{num.firm}}
}
\concept{firm functions}
