% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbase.fen.R
\name{getbase.fen}
\alias{getbase.fen}
\title{Collect information of municipality population phenomena.}
\usage{
getbase.fen(year, provincia)
}
\arguments{
\item{year}{A numerical value between 1996 and the current year indicating the year of the required data.}

\item{provincia}{one of the 52 Spain’s province.}
}
\value{
a \code{xlsx} file saved in \code{data_poblacion} folder into the working directory wtih default name \code{fen_year_provincia}.
}
\description{
Download the population phenomena data of the province`s municipalities corresponding to the required.
}
\examples{
getbase.fen(2005,”Madrid”)

}
