% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbase.fen.R
\name{getbase.fen}
\alias{getbase.fen}
\title{Collect information of municipality population phenomena.}
\usage{
getbase.fen(year, provincia)
}
\arguments{
\item{year}{A numerical value between 1996 and the current year indicating the year of the required data.}

\item{provincia}{one of the 52 Spain’s province.}
}
\value{
It is a \code{xlsx} which is host in \code{data_poblacion} folder iwhich, in turn, is located inside the working directory called \code{fen_year_provincia}.
}
\description{
\code{getbase.fen} downloads the population phenomena data of the province‘s municipalities corresponding to the required.
}
\details{
The names of the Spanish provinces may or may not be called. If yes, employ capital letters as follows:

"ALBACETE", "ALICANTE", "ALMERIA", "ARABA", "ASTURIAS", "AVILA", "BADAJOZ", "BALEARES", "BARCELONA", "BIZKAIA", "BURGOS", "CACERES", "CADIZ", "CANTABRIA", "CASTELLON", "CIUDAD REAL", "CORDOBA", "A CORUÑA", "CUENCA", "GIPUZKOA", "GIRONA", "GRANADA", "GUADALAJARA", "HUELVA", "HUESCA", "JAEN", "LEON", "LLEIDA", "LUGO", "MADRID", "MALAGA", "MURCIA", "NAVARRA", "OURENSE", "PALENCIA",  "LAS PALMAS", "PONTEVEDRA", "LA RIOJA", "SALAMANCA", "TENERIFE", "SEGOVIA", "SEVILLA","SORIA", "TARRAGONA", "TERUEL", "TOLEDO", "VALENCIA", "VALLADOLID", "ZAMORA", "ZARAGOZA", "CEUTA" and "MELILLA"
}
\examples{
getbase.fen(2005,"Madrid")

}
\seealso{
Other download functions: \code{\link{getbase.paro}},
  \code{\link{getbase.pob}}
}
\concept{download functions}
