% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbase.paro.R
\name{getbase.paro}
\alias{getbase.paro}
\title{Collect information of municipality unemploymet.}
\usage{
getbase.paro(year, mes, provincia)
}
\arguments{
\item{year}{A numerical value between 2005 and the current year indicating the year of the required data.}

\item{mes}{one of the 12 moths in Spanish indicating the moth when the data was collected.}

\item{provincia}{one of the 52 Spainish province.}
}
\value{
a file host in \code{data_paro} folder into the working directory wtih name \code{paro_MUNI_provincia_mmyy.xls}.
}
\description{
Download the work force data of the province`s municipalities corresponding to the required.
}
\details{
the possible mes value are: "enero", "febrero", "marzo", "abril", "mayo", "junio", "julio", "agosto", "septiembre", "octubre", "noviembre" and "diciembre"

You can use in capital lletter or not the names of spanish provincies. The full list of spanish provincies is:

"ALBACETE", "ALICANTE", "ALMERIA", "ARABA", "ASTURIAS", "AVILA", "BADAJOZ", " BALEARES", "BARCELONA", "BIZKAIA", "BURGOS", "CACERES", "CADIZ", "CANTABRIA", "CASTELLON", "CIUDAD REAL", "CORDOBA", "A CORUÑA", "CUENCA", "GIPUZKOA", "GIRONA", "GRANADA", "GUADALAJARA", "HUELVA", "HUESCA", "JAEN", "LEON", "LLEIDA", "LUGO", "MADRID", "MALAGA", "MURCIA", "NAVARRA", "OURENSE", "PALENCIA",  "LAS PALMAS", "PONTEVEDRA", "LA RIOJA", "SALAMANCA", "TENERIFE", "SEGOVIA", "SEVILLA","SORIA", "TARRAGONA", "TERUEL", "TOLEDO", "VALENCIA", "VALLADOLID", "ZAMORA", "ZARAGOZA", "CEUTA" and "MELILLA"
}
\examples{
getbase.paro(2005,"julio","Madrid")

}
\seealso{
Other download functions: \code{\link{getbase.fen}},
  \code{\link{getbase.pob}}
}
