% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbase.paro.R
\name{getbase.paro}
\alias{getbase.paro}
\title{Collect information of municipality unemploymet.}
\usage{
getbase.paro(year, mes, provincia)
}
\arguments{
\item{year}{A numerical value between 2005 and the current year indicating the year of the required data.}

\item{mes}{one of the 12 moths in Spanish indicating the moth when the data was collected.}

\item{provincia}{one of the 52 Spain’s province.}
}
\value{
a file host in \code{data_paro} folder into the working directory wtih name \code{paro_MUNI_provincia_mmyy.xls}.
}
\description{
Download the work force data of the province`s municipalities corresponding to the required.
}
\examples{
getbase.paro(2005,”julio”,”Madrid”)

}
