% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get required data from the PPT API}
\usage{
getData(
  type = NULL,
  cookie = NULL,
  debug = 0,
  salaries = NULL,
  keep = FALSE,
  age = 7,
  path = "//dcnsbiona01a/BIODataSVC/IN/MSP/PowerBI-Projects/dataSPA/"
)
}
\arguments{
\item{type}{the type of data that is wished to be extracted
(either \code{om} or \code{salary})}

\item{cookie}{a sessionid and csrftoken from a Department of
Fisheries and Oceans Canada (DFO) employee in the following
format: csrftoken=YOURTOKEN; sessionid=YOURSESSIONID}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}

\item{salaries}{a built in data frame obtained from the \code{dataSPA}
package that contains information about pay levels for
different DFO positions}

\item{keep}{logical value to optionally keep \code{om} data on the hard disk. Default is FALSE to no save data. Value of TRUE will save \code{om} to disk in \code{path} if the file was created more recently than \code{age} (number of days)}

\item{age}{maximum age in number of days that a file may be loaded.}

\item{path}{path to save file. Default is in the shared IN folder}
}
\value{
A) If \code{type} is \code{om}: a data frame containing the following
information: project_id, category_display, project_year_id,        "amount"
funding_source_display, id, category_type, description, fiscal_year,
project_title, status,overview, objectives, deliverables. B) If \code{type} is \code{salary}:
a data frame containing the following information: id, overtime_hours,
smart_name, duration_weeks, level_display, funding_source_display,
employee_type_display, project_year_id, project_id, fiscal_year,
project_title, median_salary, salary_per_week, amount_week,
amount_overtime, amount_total.
}
\description{
If \code{type} is \code{om}, this function extracts data from the
om-costs project planning tool (PPT) API (http://dmapps/api/ppt/om-costs/)
and combines it with the fiscal year from the project_years API
(http://dmapps/api/ppt/). If \code{type} is \code{salary}, this function
obtains the data from the staff API (http://dmapps/api/ppt/staff/)
and combines it with information from a Human Resources (HR) spreadsheet.
}
\examples{
\dontrun{
library(dataSPA)
data(salaries)
cookie <- "csrftoken=YOURTOKEN; sessionid=YOURSESSIONID"
data <- getData(type="salary", cookie=cookie, salaries=salaries)
head(data,2)
}
}
\author{
Jaimie Harbin and Remi Daigle
}
