% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEBMpillars.R
\name{getEBMpillars}
\alias{getEBMpillars}
\title{Get the discriminating words used to identify EBM pillars or objectives.}
\usage{
getEBMpillars(
  file =
    "//dcnsbiona01a/BIODataSvc/IN/MSP/PowerBI-Projects/dataSPA/inputs/Stephenson_ebfm.R"
)
}
\arguments{
\item{file}{the word file describing the EBM framwork. Defaults to the IN folder}
}
\description{
Provides words and colours to be used in highlighting.
}
\author{
Remi Daigle and Jaimie Harbin
}
