% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind.ev.R
\name{ind.ev}
\alias{ind.ev}
\title{Population index evolution}
\usage{
ind.ev(inicio, fin, provincia, print = FALSE)
}
\arguments{
\item{inicio}{starting year of the panel. Must be higher than 1996.}

\item{fin}{last year of the panel.}

\item{provincia}{one of the 52 Spanish provinces.}

\item{print}{Logical variable, do you need print a output file with the results?. Default value is FALSE}
}
\value{
a list conteining ten data frame.
}
\description{
Create a list containing the population index evolution at municipality level between the \code{inicio} and \code{fin} years.
}
\details{
in this function 10 index are comptuted:
\itemize{
\item Childhood index
	\item Youthfulness index
	\item Eld index
	\item Dependence index
	\item Total unemployment rate
	\item Male unemployment rate
	\item Femal unemployment rate
	\item Total Population average age
	\item Male population average age
\item Female population average age
}

If \code{print} is \code{TRUE}, a \code{xlsx} file containing ten sheet, one  data frame per index, is saved into the folder \code{Outputs} by the name:
\code{pob_ev_index_provincia_inicio-fin.xlsx}
}
\examples{
ind.ev(2005,2007,"Avila")

}
\references{
{
Junta de Extremadura (2017) Atlas Socieconómico de Extremadura 2017. Mérida (Spain).
\url{http://estadistica.gobex.es/web/guest/atlas-socieconomico-de-extremadura}
}
}
\seealso{
Other Manipulate functions: \code{\link{pob.e.ev}},
  \code{\link{pob.ev}}, \code{\link{pob.fen.ev}},
  \code{\link{pob.h.ev}}, \code{\link{pob.ind.p}},
  \code{\link{pob.ind}}, \code{\link{pob.m.ev}},
  \code{\link{pob.n.ev}}
}
