% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num.firm.a.R
\name{num.firm.a}
\alias{num.firm.a}
\title{Collects the set of self-employed at a municipality level}
\usage{
num.firm.a(provincia)
}
\arguments{
\item{provincia}{one of the 52 Spanish provinces. See \link{getbase.pob} for details.}
}
\value{
It is a data frame
}
\description{
\code{num.firm.a} generates a data frame with the set of self-employed in a selected municipality
}
\examples{
\dontrun{num.firm.a("Araba")}

}
\seealso{
Other firm functions: \code{\link{data.firm.a}},
  \code{\link{data.firm}}, \code{\link{num.firm}}
}
