% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num.firm.R
\name{num.firm}
\alias{num.firm}
\title{Collects the set of firms at a municipality level}
\usage{
num.firm(provincia)
}
\arguments{
\item{provincia}{one of the 52 Spanish provinces. See \link{getbase.pob} for details.}
}
\value{
It is a data frame
}
\description{
\code{num.firm} generates a data frame with the set of firms in a selected municipality
}
\examples{
\dontrun{num.firm("Araba")}

}
\seealso{
Other firm functions: \code{\link{data.firm.a}},
  \code{\link{data.firm}}, \code{\link{num.firm.a}}
}
