% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paro.R
\name{paro}
\alias{paro}
\title{Imports into R data of unemployed at a municipality level..}
\usage{
paro(year, mes = "julio", provincia)
}
\arguments{
\item{year}{a numerical value from 2005 and the latest available, which indicates the year of the required database.}

\item{mes}{one of the 12 months –in Spanish language– indicating the month of the data collection. See \link{getbase.paro} for details.}

\item{provincia}{one of the 52 Spain’s province.}
}
\value{
It is a data frame containing the fallowing variables:.
- \code{cod} is the municipality identification number based in the INE codification
- \code{Name} is the  municipality name.
- \code{Total unemployment} unemployment is the  number of unemployed in the municipality 
- \code{Total male unemployment} is the number of unemployed males in the municipality
- \code{Total female unemployment} is the number of unemployed females in the municipality
}
\description{
\code{paro} imports the unemployed of the Spanish municipalities by province.
}
\examples{
#paro(2005,"julio","Madrid")

}
\seealso{
Other Loading functions: \code{\link{pob.a}},
  \code{\link{pob.e.tot}}, \code{\link{pob.e}},
  \code{\link{pob.fen}}, \code{\link{pob.h.tot}},
  \code{\link{pob.m.tot}}, \code{\link{pob.n.ev}},
  \code{\link{pob.n.tot}}, \code{\link{pob.q}},
  \code{\link{pob.tot}}
}
