% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSPA.R
\name{plotSPA}
\alias{plotSPA}
\title{Plot dataSPA information}
\usage{
plotSPA(
  om = NULL,
  salary = NULL,
  which = NULL,
  id = NULL,
  theme = NULL,
  functionalGroup = NULL,
  section = NULL,
  division = NULL,
  region = NULL,
  item = NULL,
  funding = NULL,
  fundingChange = NULL,
  status = NULL,
  dataframe = FALSE,
  year = NULL,
  endDate = NULL,
  debug = 0
)
}
\arguments{
\item{om}{a data frame likely from \code{getData(type='om')}}

\item{salary}{a data frame likely from \code{getData(type='salary')}}

\item{which}{indicates which plot to plot (See Details).}

\item{id}{the project_id from the Project Planning Tool}

\item{theme}{theme classification of projects of either \verb{Mitigation of Ecosystem Stressors},
\verb{Marine Spatial Planning and Conservation}, \verb{Ecosystem Assessment and Climate Change},
\verb{Population Assessment and Recovery}, \code{Other}, \verb{Technology Development and Application}, and
\verb{Pacific Salmon}}

\item{functionalGroup}{classification of projects. Too see options do
\code{unique(om$function_group)}, where \code{om}is the output from \code{getData(type="om")}}

\item{section}{classification of projects referring to the sections at DFO. For
more details do \code{unique(om$section_display)}, where \code{om}is the output from
\code{getData(type="om")}}

\item{division}{classification of projects referring to the divisions at DFO. For
more details do \code{unique(om$section_display)}, where \code{om}is the output from
\code{getData(type="om")}}

\item{region}{parameter to specific specific region of either \code{Gulf}, \code{Maritimes},
\code{Pacific}, \code{Quebec}, or \verb{Ontario and Prairie}. If no region is given, a summary plot
of all regions is given.}

\item{item}{if\code{which='omAllocation'} an argument
specifying which category display to plot. If \code{which=salaryAllocation},
the argument specifys whichs level to plot. By default all are plotted}

\item{funding}{a variable used when \code{which='predictSummary'} or
\code{which='predict'} used to indicate which funding source will
experience change in the prediction. If \code{endDate} is provided,
only the length of this argument must equal 1.}

\item{fundingChange}{a percentage used when \verb{which='predictSummary}
or \code{which=predict} used to indicate by how much the funding
specified in \code{funding} will change by. If \code{endDate} is provided,
this argument is ignored.}

\item{status}{a character string indicating indicating which project
status to include (Approved,Reviewed, Draft, Submitted, Not Approved,
Recommended, and Cancelled). If NULL, all projects are included.
are included}

\item{dataframe}{a Boolean indicating if the user wants the data frame
of what is being plotted returned. For multi-year projects, a list of
data frames for each year is returned. This is mainly for package testing}

\item{year}{a character indicating which year of the project to
plot. This is given in the format "YYYY-YYYY". If \code{year} is \code{NULL}
all years are plotted by default}

\item{endDate}{the end date in "YYYY-YYYY" format that indicates
when the funding specified by \code{funding} ends. This is used
when \code{which='predict'} and can only be an interval of one year.}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}
}
\value{
None (invisible NULL).
}
\description{
This function plots specific graphs using the data
frames returned by \code{\link[=getData]{getData()}}. Note: by default, only
the approved projects are included (this can be changed
using the status argument).
}
\details{
The various plot types are as follows:
\itemize{
\item For \code{which="omBar"}, a bar graph representing amount of money ($)
coded by funding type. If plotting for a specific project, the number of
deliverables and milestones per year are shown in blue and red respectively
\item For \code{which="omPie"}, a pie chart representing amount of money ($)
coded by funding type
\item For \code{which="omAllocation"} Line graphs representing changes in O&M
investment for categories over the years. The average rate of change
per year for each category, where applicable, displayed on the graph in red
\item For \code{which=omAllocationGeneral} a pie chat representing proportion of
funding per category type is plotted for year each of the specified project
\item For \code{which="salaryBar"} a bar graph representing amount of salary money ($)
using the median values of a specific job classification. Graph color
coded by funding type indicated by the legend.
\item For \code{which="salaryAllocation"} Line graphs representing changes in Salary
investment for different classifications over the years. The average rate
of change per year for each category, where applicable, displayed on the
graph in red
\item For \code{which="weekAllocation"} a bar graph representing time (weeks)
invested per job classification for project years is plotted
\item For \code{which="indeterminate"} a bar chart representing percentage of
indeterminate vs non-Indeterminate employees for project years.
Number of staff shown on the figure in red
\item For \code{which="predictSummary"}a line chart showing the trends for
changes in different funding scenarios
\item For \code{which="predict"} a bar chart showing the trends for changes
in different funding scenarios is plotted with the number of stations
overlying the figure with a line. A new plot is plotted for each
funding scenario. If \code{endDate} is provided, the bar chart will
represent funding as status quo until the end date (See example 3)
\item For \verb{which="predictSalary} a bar chat showing the trends of salary
increasing while funding amount remains the same. Note: This plot does
not include amount of overtime as that is hard to predict.
\item For \verb{which="predictOM} a bar chart showing the trends of om
increasing while funding amoutn remains the same. This plot
assumes an inflation rate of 2\%.
\item For \code{which="overviewStatus"} a pie chart is created that shows the
number of projects for each of the following status': Approved,
Draft, Submitted, and Reviews for the specified parameters.
\item For \code{which="overviewInvestment"} a bar chart indicating how much
money was spent on A) OM investment (if an om argument is provided),
B) Salary investment (if a salary argument is provided), or C)
OM and Salary investment (if both om & salary argument given).
Note: This includes overtime salary amount.
}
}
\examples{
\dontrun{
# Example 1: Plot Bar graph of O&M Allocations
library(dataSPA)
data <- getData(type="om", cookie=cookie)
plotSPA(om=data, which="omBar")

# Example 2: Plot a prediction of number of station
# for NCP funding decreasing by 25\%, 50\%, 75\%, and 100\%
for the Snow Crab project

library(dataSPA)
data(salaries)
data <- getData(type='om', cookie=cookie)
data2 <- getData(type='salary', cookie=cookie)
plotSPA(om=data, salary=data2, which='predictSummary', id=1093,
funding= rep("NCP (A-base)",4),
fundingChange=c(-25, -5, -75, -100))

# Example 3: Bar plot predicting of the change in
number of stations impact as a result of a certain funding
ending
plotSPA(om=data, salary=data2, which='predict', id=1093,
funding= "NCP (A-base)", endDate="2026-2027")

# Example 4: Pie Chart showing number of projects by status
plotSPA(om=data, salary=data2, which="overviewStatus")

}
}
\author{
Jaimie Harbin
}
