% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSPA.R
\name{plotSPA}
\alias{plotSPA}
\title{Plot an dataSPA information}
\usage{
plotSPA(
  om = NULL,
  salary = NULL,
  which = NULL,
  id = NULL,
  funding = NULL,
  fundingChange = NULL,
  dataframe = FALSE,
  year = NULL,
  endDate = NULL,
  salaries = NULL,
  debug = 0
)
}
\arguments{
\item{om}{a data frame likely from \code{getData(type='om')}}

\item{salary}{a data frame likely from \code{getData(type='salary')}}

\item{which}{indicates which plot to plot (See Details). The
options to use include: \code{omBar}, \code{omPie}, \code{omAllocation},
\code{omAllocationGeneral}, \code{salaryBar}, \code{salaryAllocation},
\code{weekAllocation},\code{indeterminate}, \code{predictSummary},\code{predict},
or \code{predictSalary}}

\item{id}{the project_id from the Project Planning Tool}

\item{funding}{a variable used when \code{which='predictSummary'} or
\code{which='predict'} used to indicate which funding source will
experience change in the prediction. If \code{endDate} is provided,
only the length of this argument must equal 1.}

\item{fundingChange}{a percentage used when \verb{which='predictSummary}
or \code{which=predict} used to indicate by how much the funding
specified in \code{funding} will change by. If \code{endDate} is provided,
this argument is ignored.}

\item{dataframe}{a Boolean indicating if the user wants the data frame
of what is being plotted returned. For multi-year projects, a list of
data frames for each year is returned. This is mainly for package testing}

\item{year}{a character indicating which year of the project to
plot. This is given in the format "YYYY-YYYY". If \code{year} is \code{NULL}
all years are plotted by default}

\item{endDate}{the end date in "YYYY-YYYY" format that indicates
when the funding specified by \code{funding} ends. This is used
when \code{which='predict'} and can only be an interval of one year.}

\item{salaries}{a built in data frame obtained from the \code{dataSPA}
package that contains information about pay levels for
different DFO positions}

\item{debug}{integer value indicating level of debugging.
If this is less than 1, no debugging is done. Otherwise,
some functions will print debugging information.}
}
\value{
None (invisible NULL).
}
\description{
This function plots specific graphs (determined by the
\code{which} argument) using the data frames returned by
\code{\link[=getData]{getData()}}.
}
\details{
The various plot types are as follows:
\itemize{
\item For \code{which="omBar"}, a bar graph representing amount of money ($)
coded by funding type.
\item For \code{which="omPie"}, a pie chart representing amount of money ($)
coded by funding type
\item For \code{which="omAllocation"} Line graphs representing changes in O&M
investment for categories over the years for a specific project.
Average rate of change per year for each category, where applicable,
displayed on the graph in red
\item For \code{which=omAllocationGeneral} a pie chat representing proportion of
funding per category type is plotted for year each of the specified project
\item For \code{which="salaryBar"} a bar graph representing amount of salary money ($)
using the median values of a specific job classification. Graph color
coded by funding type indicated by the legend.
\item For \code{which="salaryAllocation"} Line graphs representing changes in Salary
investment for different classifications over the years for a specific project.
Average rate of change per year for each category, where applicable,
displayed on the graph in red
\item For \code{which="weekAllocation"} a bar graph representing time (weeks)
\item For \code{which="indeterminate"} a bar chart representing percentage of indeterminate vs
non-Indeterminate employees for project years. Number of staff shown on the
figure in red
\item For \code{which="predictSummary"}a line chart showing the trends for
changes in different funding scenarios
\item For \code{which="predict"} a bar chart showing the trends for changes
in different funding scenarios is plotted with the number of stations
overlying the figure with a line. A new plot is plotted for each
funding scenario. If \code{endDate} is provided, the bar chart will
represent funding as status quo until the end date (See example 3)
\item For \verb{which="predictSalary} a bar chat showing the trends of salary
increasing while funding amount remains the same. Note: This plot does
not include amount of overtime as that is hard to predict.
}
}
\examples{
\dontrun{
# Example 1: Plot Bar graph of O&M Allocations
library(dataSPA)
data <- getData(type="om", cookie=cookie)
plotSPA(om=data, which="omBar")

# Example 2: Plot a prediction of number of station
# for NCP funding decreasing by 25\%, 50\%, 75\%, and 100\%
for the Snow Crab project

library(dataSPA)
data(salaries)
data <- getData(type='om', cookie=cook
data2 <- getData(type='salary', cookie=cookie)
plotSPA(om=data, salary=data2, which='predictSummary', id=1093,
funding= rep("NCP (A-base)",4),
fundingChange=c(-25, -5, -75, -100), salaries=salaries)

# Example 3: Bar plot predicting of the change in
number of stations impact as a result of a certain funding
ending

plotSPA(om=data, salary=data2, which='predict', id=1093,
funding= "NCP (A-base)", endDate="2026-2027", salaries=salaries)

}
}
\author{
Jaimie Harbin
}
