% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.a.R
\name{pob.a}
\alias{pob.a}
\title{Population grouped by age data}
\usage{
pob.a(year, provincia)
}
\arguments{
\item{year}{a numerical value from 1996 and the last available, which indicates the year of the required database.}

\item{provincia}{one of the 52 Spainish provinces.}
}
\value{
It is a list containing a total population data frame and the population grouped by sex. Each data frame contains the following variables:: 

\itemize{
		\item \code{cod} is the municipality identification number based in the INE codification.
	\item \code{Name} the municipality name.
		\item \code{Total} total municipality population
		\item A hundred and one variables containing the population grouped by one-year age
}
}
\description{
\code{pob.a} imports into an R file the municipality population database grouped by age
}
\examples{
pob.a(2016,"Caceres")

}
\seealso{
Other Loading functions: \code{\link{paro}},
  \code{\link{pob.e.tot}}, \code{\link{pob.e}},
  \code{\link{pob.fen}}, \code{\link{pob.h.tot}},
  \code{\link{pob.m.tot}}, \code{\link{pob.n.ev}},
  \code{\link{pob.n.tot}}, \code{\link{pob.q}},
  \code{\link{pob.tot}}
}
