% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.a.R
\name{pob.a}
\alias{pob.a}
\title{Population grouped by age data}
\usage{
pob.a(year, provincia)
}
\arguments{
\item{year}{A numerical value between 1996 and the current year indicating the year of the required data.}

\item{provincia}{one of the 52 Spainish province.}
}
\value{
a list containing a total population data frame and the population grouped by sex. Each data frame contains the following variables: 

\itemize{
		\item \code{cod} is the municipality identification number based in the INE codification.
	\item \code{Name} the municipality name.
		\item \code{Total} total municipality population
		\item  101 variables containing the population grouped by age
}
}
\description{
import into R the  population grouped by age data
}
\examples{
pob.a(2016,"Caceres")

}
\seealso{
Other Loading functions: \code{\link{paro}},
  \code{\link{pob.e.tot}}, \code{\link{pob.e}},
  \code{\link{pob.fen}}, \code{\link{pob.h.tot}},
  \code{\link{pob.m.tot}}, \code{\link{pob.n.ev}},
  \code{\link{pob.n.tot}}, \code{\link{pob.q}},
  \code{\link{pob.tot}}
}
