% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.e.ev.R
\name{pob.e.ev}
\alias{pob.e.ev}
\title{Total foreign population evolution}
\usage{
pob.e.ev(inicio, fin, provincia, print = FALSE)
}
\arguments{
\item{inicio}{starting year of the panel. Must be higher than 1996.}

\item{fin}{last year of the panel.}

\item{provincia}{one of the 52 Spanish provinces.}

\item{print}{Logical variable, do you need print a output file with the results?. Default value is FALSE}
}
\value{
a data frame
}
\description{
Create a data drame containing the provincie total foreign population evolution at municipality level between the \code{inicio} and \code{fin} years.
}
\details{
If \code{print} is \code{TRUE}, a \code{xlsx} file containing the data frame is saved into the folder \code{Outputs} by the name:
\code{pob_foreign_ev_provincia_inicio-fin.xlsx}
}
\examples{
pob.ev(2005,2007,"Avila")

}
\seealso{
Other Manipulate functions: \code{\link{ind.ev}},
  \code{\link{pob.ev}}, \code{\link{pob.fen.ev}},
  \code{\link{pob.h.ev}}, \code{\link{pob.ind.p}},
  \code{\link{pob.ind}}, \code{\link{pob.m.ev}},
  \code{\link{pob.n.ev}}
}
