% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.e.ev.R
\name{pob.e.ev}
\alias{pob.e.ev}
\title{Panel of foreign population at a municipality level for a period of time}
\usage{
pob.e.ev(inicio, fin, provincia, print = FALSE)
}
\arguments{
\item{inicio}{starting year of the panel, which must be higher than 1996.}

\item{fin}{last year of the panel.}

\item{provincia}{one of the 52 Spanish provinces.}

\item{print}{logical variable ‘do you need print a output file with the results?’ for which FALSE is the default value}
}
\value{
It is a data frame
}
\description{
\code{pob.e.ev} creates a data frame with a panel of  total foreign population at a municipality level for a period of time from the  \code{inicio} to \code{fin}.
}
\details{
If \code{print} is set to \code{TRUE}, an \code{xlsx} file containing the data frame is saved int the folder \code{Outputs} called \code{pob_foreign_ev_provincia_inicio-fin.xlsx}
}
\examples{
pob.ev(2005,2007,"Avila")

}
\seealso{
Other Manipulate functions: \code{\link{pob.ev}},
  \code{\link{pob.fen.ev}}, \code{\link{pob.h.ev}},
  \code{\link{pob.ind.p}}, \code{\link{pob.ind}},
  \code{\link{pob.m.ev}}, \code{\link{pob.n.ev}}
}
\concept{Manipulate functions}
