% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.e.R
\name{pob.e}
\alias{pob.e}
\title{Population grouped by nationality and sex data}
\usage{
pob.e(year, provincia)
}
\arguments{
\item{year}{A numerical value between 1996 and the current year indicating the year of the required data.}

\item{provincia}{one of the 52 Spain`s province.}
}
\value{
a list containing a total population data frame and the population grouped by sex. Each data frame contains the following variables: 
\itemize{
		\item \code{cod} is the municipality identification number based in the INE codification.
	\item \code{Name} the municipality name.
		\item \code{Total} total municipality population
		\item Three variables containing the population divided into Younger than 16 years old, Between 16 and 64 years old and Older than 65 years old grouped by total population, foreign population and national population.
}
}
\description{
import into R the  population grouped by nationality and sex data
}
