% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.e.tot.R
\name{pob.e.tot}
\alias{pob.e.tot}
\title{Total Foreign Population}
\usage{
pob.e.tot(year, provincia)
}
\arguments{
\item{year}{a numerical value from 1996 and the latest available year, which indicates the year of the required database.}

\item{provincia}{one of the 52 Spainish provinces.}
}
\value{
It is a data frame containing the municipality total foreign population
}
\description{
\code{pob.e.tot} imports into an R file the municipality total foreign population.
}
\examples{
pob.e.tot(2016,"Madrid")

}
\seealso{
Other Loading functions: \code{\link{paro}},
  \code{\link{pob.a}}, \code{\link{pob.e}},
  \code{\link{pob.fen}}, \code{\link{pob.h.tot}},
  \code{\link{pob.m.tot}}, \code{\link{pob.n.ev}},
  \code{\link{pob.n.tot}}, \code{\link{pob.q}},
  \code{\link{pob.tot}}
}
