% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.e.tot.R
\name{pob.e.tot}
\alias{pob.e.tot}
\title{Total Foreign Population  data}
\usage{
pob.e.tot(year, provincia)
}
\arguments{
\item{year}{A numerical value between 1996 and the current year indicating the year of the required data.}

\item{provincia}{one of the 52 Spain`s province.}
}
\value{
a  data frame conatining the municipality total foreign population
}
\description{
import into R the total foreign municpality population
}
