% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.fen.R
\name{pob.fen}
\alias{pob.fen}
\title{Principal municiplaity puplation phenomena}
\usage{
pob.fen(year, provincia)
}
\arguments{
\item{year}{A numerical value between 1996 and 2015 year indicating the year of the required data.}

\item{provincia}{one of the 52 Spainish province.}
}
\value{
a  data frame conatining the principal population phenomena 
\itemize{
	\item \code{Birth} number of birth in the municipality
	\item \code{Deaths} Number of death in the municipality 
}
}
\description{
import into R the principal population phenomena
}
\examples{
pob.fen(2012,"Madrid")

}
\seealso{
Other Loading functions: \code{\link{paro}},
  \code{\link{pob.a}}, \code{\link{pob.e.tot}},
  \code{\link{pob.e}}, \code{\link{pob.h.tot}},
  \code{\link{pob.m.tot}}, \code{\link{pob.n.ev}},
  \code{\link{pob.n.tot}}, \code{\link{pob.q}},
  \code{\link{pob.tot}}
}
