% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.fen.R
\name{pob.fen}
\alias{pob.fen}
\title{Number of births and deaths at a municipality level}
\usage{
pob.fen(year, provincia)
}
\arguments{
\item{year}{a numerical value from 1996 and 2015, which indicates the year of the required database}

\item{provincia}{one of the 52 Spainish province.}
}
\value{
It is a data frame containing two main demographic indexes: 
\itemize{
	\item \code{Birth} number of birth in the municipality
	\item \code{Deaths} Number of death in the municipality 
}
}
\description{
\code{pob.fen} imports into an R file two main demographic indexes at the municipality level: number of births and deaths.
}
\examples{
pob.fen(2012,"Madrid")

}
\seealso{
Other Loading functions: \code{\link{paro}},
  \code{\link{pob.a}}, \code{\link{pob.e.tot}},
  \code{\link{pob.e}}, \code{\link{pob.h.tot}},
  \code{\link{pob.m.tot}}, \code{\link{pob.n.ev}},
  \code{\link{pob.n.tot}}, \code{\link{pob.q}},
  \code{\link{pob.tot}}
}
