% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.ind.p.R
\name{pob.ind.p}
\alias{pob.ind.p}
\title{Calculate the population indexs}
\usage{
pob.ind.p(year, provincia, print = FALSE)
}
\arguments{
\item{year}{A numerical value between 1996 and the current year indicating the year of the required data.}

\item{provincia}{one of the 52 Spanish provinces.}

\item{print}{Logical variable, do you need print a output file with the results?. Default value is FALSE}
}
\value{
a data frame
}
\description{
Compute a set of population index
}
\details{
in this function 10 index are comptuted:
\itemize{
\item Childhood index
	\item Youthfulness index
	\item Eld index
	\item Dependence index
	\item Total unemployment rate
	\item Male unemployment rate
	\item Femal unemployment rate
	\item Total Population average age
	\item Male population average age
\item Female population average age
}

The principal difference between this function and \code{pob.ind()} is the precision of the calculaton. This function use since the year 2011 the exact age definition for index imputation, instead that the aproximation use in \code{pob.ind}. For more details see the 2017 Socioeconomic Atlas of Extremadura.
}
\examples{
pob.ind.p(2012,"Madrid")

}
\references{
{
Junta de Extremadura (2017) Atlas Socieconómico de Extremadura 2017. Mérida (Spain).
\url{http://estadistica.gobex.es/web/guest/atlas-socieconomico-de-extremadura}
}
}
\seealso{
Other Manipulate functions: \code{\link{ind.ev}},
  \code{\link{pob.e.ev}}, \code{\link{pob.ev}},
  \code{\link{pob.fen.ev}}, \code{\link{pob.h.ev}},
  \code{\link{pob.ind}}, \code{\link{pob.m.ev}},
  \code{\link{pob.n.ev}}
}
