% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.m.tot.R
\name{pob.m.tot}
\alias{pob.m.tot}
\title{Female Population  data}
\usage{
pob.m.tot(year, provincia)
}
\arguments{
\item{year}{a numerical value from 1996 and the latest available year, which indicates the year of the required database.}

\item{provincia}{one of the 52 Spainish provinces.}
}
\value{
It is a  data frame conatining the municipality total female population
}
\description{
\code{pob.m.tot} imports into an R file the female population of Spain by municipality.
}
\examples{
pob.m.tot(2016,"Madrid")

}
\seealso{
Other Loading functions: \code{\link{paro}},
  \code{\link{pob.a}}, \code{\link{pob.e.tot}},
  \code{\link{pob.e}}, \code{\link{pob.fen}},
  \code{\link{pob.h.tot}}, \code{\link{pob.n.ev}},
  \code{\link{pob.n.tot}}, \code{\link{pob.q}},
  \code{\link{pob.tot}}
}
\concept{Loading functions}
