% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.n.ev.R
\name{pob.n.ev}
\alias{pob.n.ev}
\title{Panel of the national population at the municipality level for a period of time}
\usage{
pob.n.ev(inicio, fin, provincia, print = FALSE)
}
\arguments{
\item{inicio}{starting year of the panel, which must be higher than 1996}

\item{fin}{last year of the panel.}

\item{provincia}{one of the 52 Spanish provinces.}

\item{print}{logical variable ‘do you need to print a output file with the results?’ being FALSE the default value}
}
\value{
It is a data frame
}
\description{
\code{pob.n.ev} creates a data frame containing the panel of the national population of Spain at the municipality level for a period of time from the years  \code{inicio} to \code{fin}.
}
\details{
If \code{print} is set to \code{TRUE}, an \code{xlsx} file containing the data frame is saved in the folder \code{Outputs} which name \code{pob_national_ev_provincia_inicio-fin.xlsx}
}
\examples{
pob.n.ev(2005,2007,"Avila")

}
\seealso{
Other Loading functions: \code{\link{paro}},
  \code{\link{pob.a}}, \code{\link{pob.e.tot}},
  \code{\link{pob.e}}, \code{\link{pob.fen}},
  \code{\link{pob.h.tot}}, \code{\link{pob.m.tot}},
  \code{\link{pob.n.tot}}, \code{\link{pob.q}},
  \code{\link{pob.tot}}

Other Manipulate functions: \code{\link{pob.e.ev}},
  \code{\link{pob.ev}}, \code{\link{pob.fen.ev}},
  \code{\link{pob.h.ev}}, \code{\link{pob.ind.p}},
  \code{\link{pob.ind}}, \code{\link{pob.m.ev}}
}
