% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.q.R
\name{pob.q}
\alias{pob.q}
\title{Population by five-year age groups and sex}
\usage{
pob.q(year, provincia)
}
\arguments{
\item{year}{a numerical value from 1996 and the latest available year, which indicates the year of the required database.}

\item{provincia}{one of the 52 Spainish provinces.}
}
\value{
It is a list containing a data frame of the Spanish municipality population by five-year age groups and sex. Each data frame contains the following variables:: 
- \code{cod} is the municipality identification number based in the INE codification.
- \code{Name} is the municipality name.
- \code{Total} is the municipality population.
- Twenty-one variables containing the population by five-year age groups.
}
\description{
\code{pob.q} imports into an R file the Spanish municipality population by five-year age groups and sex
}
\examples{
pob.q(2016,"Caceres")

}
\seealso{
Other Loading functions: \code{\link{paro}},
  \code{\link{pob.a}}, \code{\link{pob.e.tot}},
  \code{\link{pob.e}}, \code{\link{pob.fen}},
  \code{\link{pob.h.tot}}, \code{\link{pob.m.tot}},
  \code{\link{pob.n.ev}}, \code{\link{pob.n.tot}},
  \code{\link{pob.tot}}
}
