% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob.tot.R
\name{pob.tot}
\alias{pob.tot}
\title{Population  data}
\usage{
pob.tot(year, provincia)
}
\arguments{
\item{year}{a numerical value from 1996 and the latest available year, which indicates the year of the required database}

\item{provincia}{one of the 52 Spainish provinces.}
}
\value{
It is a data frame containing the total population of Spain at the municipality level.
}
\description{
\code{pob.tot} imports into an R file the total population of Spain by municipality.
}
\examples{
pob.tot(2016,"Madrid")

}
\seealso{
Other Loading functions: \code{\link{paro}},
  \code{\link{pob.a}}, \code{\link{pob.e.tot}},
  \code{\link{pob.e}}, \code{\link{pob.fen}},
  \code{\link{pob.h.tot}}, \code{\link{pob.m.tot}},
  \code{\link{pob.n.ev}}, \code{\link{pob.n.tot}},
  \code{\link{pob.q}}
}
\concept{Loading functions}
