% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetSPA.R
\name{subsetSPA}
\alias{subsetSPA}
\title{Subset O&M and salary data frames}
\usage{
subsetSPA(
  om = NULL,
  salary = NULL,
  approved = TRUE,
  region = NULL,
  theme = NULL,
  functionalGroup = NULL,
  section = NULL,
  division = NULL
)
}
\arguments{
\item{om}{a data frame likely from \code{getData(type='om')}}

\item{salary}{a data frame likely from \code{getData(type='salary')}}

\item{approved}{a boolean indicating if the plots should only include
the approved projects. If FALSE, projects of all status (Approved,
Reviewed, Draft, Submitted, Not Approved, Recommended, and Canceled)
are included}

\item{region}{parameter to specific specific region of either \code{Gulf}, \code{Maritimes},
\code{Pacific}, \code{Quebec}, or \verb{Ontario and Prairie}.}

\item{theme}{theme classification of projects. This could be set to \code{return}
to see a list of possibilities. See examples below.}

\item{functionalGroup}{classification of projects. This could be set to \code{return}
to see a list of possibilities. See examples below.}

\item{section}{classification of projects referring to the sections at DFO.
This could be set to \code{return} to see a list of possibilities. See examples below.}

\item{division}{classification of projects referring to the divisions at DFO.
This could be set to \code{return} to see a list of possibilities. See examples below.}
}
\description{
This function either A) retusn a subset data frame of O&M or salary
data frame by region, approved projects, theme,
functional group, section, or division or B)
returns a list of theme, functional group, section,
or division of the setset if either of the above
mentioned arguments are set to "return"
}
\examples{
\dontrun{
# Example 1: Subset for approved projects in the Maritimes Region
# for the Coastal Exosystem Science Division
library(dataSPA)
om <- getData(type="om", cookie=cookie, age=100)
d <- subsetSPA(om=om, approved=TRUE, region="Maritimes",
division ="Coastal Ecosystem Science Division")
head(d)

# Example 2: Get a list of the sections that contain approved projects
# in the Maritimes Region
d <- subsetSPA(om=om, approved=TRUE, region="Maritimes",
section ="return")
}
}
